+++ Interface/AddOns/nibRealUI/Core.lua
@@ -92,16 +92,16 @@ RealUI.defaultPositions = {
         ["HuDX"] = 0,
         ["HuDY"] = -38,
         ["UFHorizontal"] = 200,
-        ["ActionBarsY"] = -116,
+        ["ActionBarsY"] = -115.5,
         ["ActionBarsBotY"] = 16,
         ["GridTopX"] = 0,
         ["GridTopY"] = -197.5,
         ["GridBottomX"] = 0,
         ["GridBottomY"] = 58,
         ["CastBarPlayerX"] = 0,
-        ["CastBarPlayerY"] = 0,
+        ["CastBarPlayerY"] = -20,
         ["CastBarTargetX"] = 0,
-        ["CastBarTargetY"] = 0,
+        ["CastBarTargetY"] = -20,
         ["SpellAlertWidth"] = 150,
         ["BossX"] = -32,        -- Boss anchored to RIGHT
         ["BossY"] = 314,
@@ -137,7 +137,7 @@ local defaults, charInit do
     }
     local spec = {}
     for specIndex = 1, RealUI.numSpecs do
-        local _, _, _, _, _, role = _G.GetSpecializationInfoForClassID(RealUI.classID, specIndex)
+        local _, _, _, _, role = _G.GetSpecializationInfoForClassID(RealUI.classID, specIndex)
         debug("Spec info", specIndex, role)
         spec[specIndex] = role == "HEALER" and 2 or 1
     end