+++ Interface/AddOns/nibRealUI/Libs/Fixes/Taint.lua
@@ -1,19 +1,4 @@
-local ADDON_NAME, private = ...
-
 do --[[ World Map ]]--
-    -- original code by ls- (lightspark)
-    local old_ResetZoom = _G.WorldMapScrollFrame_ResetZoom
-    _G.WorldMapScrollFrame_ResetZoom = function()
-        if _G.InCombatLockdown() then
-            _G.WorldMapFrame_Update()
-            _G.WorldMapScrollFrame_ReanchorQuestPOIs()
-            _G.WorldMapFrame_ResetPOIHitTranslations()
-            _G.WorldMapBlobFrame_DelayedUpdateBlobs()
-        else
-            old_ResetZoom()
-        end
-    end
-
     local old_QuestMapFrame_OpenToQuestDetails = _G.QuestMapFrame_OpenToQuestDetails
     _G.QuestMapFrame_OpenToQuestDetails = function(questID)
         if _G.InCombatLockdown() then
@@ -27,28 +12,11 @@ do --[[ World Map ]]--
 
     _G.WorldMapFrame.questLogMode = true
     _G.QuestMapFrame_Open(true)
-end
 
 
-do --[[ Artifact Frame ]]--
-    -- original code by Gnarfoz
-    --C_ArtifactUI.GetTotalPurchasedRanks() shenanigans
-    local oldOnShow
-    local newOnShow
-
-    local function newOnShow(self)
-        if C_ArtifactUI.GetTotalPurchasedRanks() then 
-            oldOnShow(self)
-        else
-            HideUIPanel(ArtifactFrame)
-        end
-    end
-
-    local function artifactHook()
-        if not oldOnShow then
-            oldOnShow = ArtifactFrame:GetScript("OnShow")
-            ArtifactFrame:SetScript("OnShow", newOnShow)
-        end
-    end
-    hooksecurefunc("ArtifactFrame_LoadUI", artifactHook)
+    local old_WorldMapFrame_OnHide = _G.WorldMapFrame:GetScript("OnHide")
+    _G.WorldMapFrame:SetScript("OnHide", function(self)
+        if _G.InCombatLockdown() then return end
+        old_WorldMapFrame_OnHide(self)
+    end)
 end